carla-1/standalone_carla_916/requirements.txt
# Shenron Standalone Agent - Requirements
# For Python 3.12 + CARLA 0.9.16

carla==0.9.16
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.10.0
filterpy>=1.4.5
pynvml>=11.5.0
mat4py>=0.5.0
matplotlib>=3.7.0
ujson>=5.0.0
Pillow>=9.0.0
PyYAML>=6.0
